requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
pandas==2.0.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
python-dotenv==1.0.0
